Default weight matrix

Weights are given for each pair of well name (pumping well PW or observation wells) and name of goodness of fit (gof)-function.

ucDefaultConfWeights(
  wells = wtConfiguredWellnames(wtDefaultConfiguration()),
  gofs = c("RMSE", "NSE", "mNSE")
)

Arguments

wells

well names, default: wtConfiguredWellnames(wtDefaultConfiguration())

gofs

default: c("RMSE", "NSE", "mNSE")

Value

Matrix of weights used when calculating overall fitness. Row names = names of fitness indicators out of the set of c("ME", "MAE", "MSE", "RMSE", "NRMSE", "PBIAS", "RSR", "rSD", "NSE", "mNSE", "rNSE", "d", "md", "rd", "cp", "r", "R2", "bR2", "KGE, "VE"). Column names = well names ("PW" = pumping well, observation wells named as in parameter setting).

See also