Set placeholder in WTAQ parameter setting

Set placeholder for parameter value in WTAQ parameter setting

ucSetPlaceholder(parname, wtaqConfiguration, pchar = "@")

Arguments

parname
Parameter name. Must be one of the aquifer parameters "bb", "hkr", "hkz", "ss", "sy" (for description see wtConfigureAquifer) or one of the drainage parameters "acc", "akk", "amm", "axmm" (for description see wtConfigureDrainage; parameter "alpha" currently not supported) or one of the pumpwell parameters "qq", "rw", "rc", "zpd", "zpl", "sw" (for description see wtConfigurePumpwell) or one of the observation well parameters "r_i", "z1_i", "z2_i", "zp_i", "rp_i", "xll_i" (for description see wtConfigureObservationWell) where "i" is the number of the observation well according to the list index in the "obswell" section of the parameter setting.

wtaqConfiguration
WTAQ parameter setting as created by means of wtConfigure

pchar
character to be used as the very first and very last character of the placeholder. Default: “@”

Value

Parameter setting in which a parameter value is replaced by an appropriate placeholder

See also